refactor: separate functions into different files
